html/semantics/interactive-elements/the-dialog-element/close-event-microtask.tentative.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/whatwg/html/issues/9046">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<dialog>dialog</dialog>

<script>
promise_test(async () => {
  const dialog = document.querySelector('dialog');
  dialog.showModal();

  let gotClose = false;
  dialog.addEventListener('close', () => gotClose = true);

  // Asserting inside the microtask callback causes a timeout, so pull the
  // value out instead.
  const gotCloseInTime = await new Promise(resolve => {
    dialog.close();
    queueMicrotask(() => {
      resolve(gotClose);
    });
  });
  assert_true(gotCloseInTime);
}, '<dialog> close events should be fired at microtask timing.');
</script>
